+++ R/readClusterDesc.R
@@ -104,18 +104,35 @@ readClusterSTDesc <- function(opts = simOptions()) {
     "st-storage/clusters" =  "st-storages"
   )
   
-  if(api_study){
-    
+  if (api_study) {
     # api request with all columns
-    list_clusters = api_get(
+    list_clusters <- api_get(
       opts = opts,
       endpoint = paste0(opts$study_id, "/table-mode/", table_type),
-      query = list(
-        columns = ""
-      )
+      query = list(columns = "")
     )
     
-    return(list_clusters)
+    if (length(list_clusters) == 0) {
+      return(data.table())
+    } 
+    
+    rows_cluster <- lapply(names(list_clusters), FUN = function(cl) {
+        splitted_name <- unlist(strsplit(cl, split = " / "))
+        row_cluster <- as.data.frame(list_clusters[[cl]])
+        row_cluster$area <- splitted_name[1]
+        row_cluster$cluster <- splitted_name[2]
+  
+        return(row_cluster)
+      }
+    )
+
+    df_clusters <- do.call("rbind", rows_cluster)
+    id_cols <- intersect(c("area", "cluster", "group"), colnames(df_clusters))
+    additional_cols <- setdiff(colnames(df_clusters), id_cols)  
+    df_clusters <- df_clusters[,c(id_cols, additional_cols)]
+    df_clusters$cluster <- as.factor(tolower(df_clusters$cluster))
+    
+    return(as.data.table(df_clusters))
   }
       
   # "text" mode
@@ -124,7 +141,7 @@ readClusterSTDesc <- function(opts = simOptions()) {
   # READ cluster properties
   properties <- get_input_cluster_properties(table_type = table_type, 
                                              opts = opts)
-  
+   
   # read properties for each area
   res <- plyr::llply(areas, function(x, prop_ref=properties) {
     clusters <- readIniFile(file.path(path, x, "list.ini"))